Add tests for offline content conversion
This adds a test case for all offline database formats to show that
they correctly convert to the current internal manifest format.

Because the database dumps were made with a buggy version of Shaka
Player, the MIME types in the various periods were not consistent.
This corrects the MIME types and codecs in the database dumps to
video/webm and vp9.

Change-Id: I0cf7ddead8352030e87474de6fed15a3ac1bb553