Allows VTT files with erroneous linebreaks
The VTT spec does not allow more than
one linebreak within the same cue. It's
desirable to still allow most of the VTT file
to work even if the file is technically out of
spec, however. This commit allows such
files by removing a thrown exception in
the case of bad time codes, instead
returning null, skipping the erroneous cue.
The text_parser also needed to be
patched, as it previously threw a perhaps
unexpected exception when data_.length
is accessed on a null or undefined data
object.