Fix argument count on compiled default callback
This fixes the argument count on the compiled versions of our default
callbacks.  This makes sure we don't issue incorrect warnings about
the application's configured callbacks.

This was discovered in the compiled build of our own demo by observing
the warning logs about the wrong number of arguments in some of our
callbacks.

Backported to v2.5.x

Change-Id: I401d54bdc706aee9a70fbf5db83fc5e7de5cb183