Convert text track visibility to bool
This is not a real bug per se, but the demo to reproduce issue #2399
would sometimes produce 1 from isTextTrackVisible() because the demo
called setTextTrackVisibility(1) instead of
setTextTrackVisibility(true).  Internally, we should convert the input
to a boolean so that we can always return the correct type from
isTextTrackVisible().

Backported to v2.5.x

Change-Id: I9d3e4b6b6035071819256b1130693d998cb3e20c