Add config to show an unbuffered at seek bar start
If set, this new config field will show any unbuffered region at the
start of the seek bar.  This matches the v2.5 default behavior.  If
unset (the v2.6 default), this will paint any unbuffered region at the
start of the seek bar as simply "played" content (matching Chrome's
native seek bar).

Closes #2424

Backported to v2.5.x, with the default set to true in this branch.

Change-Id: I528169982ce493b0dbf5a8a1bb46e9e096f422d6