Remove the "Unknown" source.
No asset was using "Unknown" as a source, and it was taking up
space in the search drop-downs and being confusing. It was only used
when loading unknown assets from the URL... but then, the source of
an asset loaded from the URL is more or less irrelevant, so that
doesn't really matter.

Change-Id: I1e20c717351aaf66f8a2b3fa25422ed884c7fd80