Fix adaptation event tests in uncompiled mode
These tests failed an assertion because they did not respect the rules
for AbrManager.  In compiled mode, this went unnoticed because asserts
are removed in the compiled bundle.  But when running tests in
uncompiled mode, these would always fail the assertion and therefore
fail the test.

Backported to v2.5.x

Change-Id: I44700d118f3f45ebcc955d9306deaca4815b902e