Don't cache everything from the demo origin in service worker
Before this change, location.origin appeared in a list of cacheable
URL prefixes in the service worker.  This was a lazy way of making
sure that all node modules, translations, etc. would be cached without
being explicitly listed.  But it also meant that we could wind up
caching the uncompiled sources, which can complicate debugging if they
are inadvertently pulled from cache on reload.

This change adds the missing local resources that can't be listed
explicitly (translations, logos) by using absolute URL prefixes.

Change-Id: I4963516ec392e7607f86bc891bcdcf46c141951d