Resize ad after 'loadeddata' event.
Before, we depended on the 'loadedmetadata' event to know when to resize
an ad. However, on Safari, that event would sometimes fire before the
video element had changed to its final size. This would result in the ad
being the wrong size.
This changes the ad manager to instead wait for the 'loadeddata' event,
which happens later in the loading process.

Closes #2462

Change-Id: Ie0470eb894034e1427a0767c771a01f801d0f9b4