Fix test flake in very large MPD test
While fixing cherry-picks for v2.5.x, I found that the test for
parsing very large MPDs would fail on that branch because it referred
to segments that do not exist locally.  This changes the MPD to refer
to local segments which can be loaded, which is important on v2.5.x
where the load() call can fail if the first segment fails.

This tweak makes the test work universally on both branches.

Change-Id: I97284aa964034bf5768abdb97f06b2a1e371d865