v2.5.x search hash parameters compatibility
When backporting the search parameters in hash feature to v2.5.x,
the mechanism behind it was modified to handle that branch's lack
of localized strings.
This updates the master version of the feature to follow that change;
now, instead of storing the value of the feature, we store the key.

Change-Id: Ic18aa2ff3d138a68cb3acc154aeb8c7e992daef3