Fix UI failure on destroy
When destroying the UI, it destroys the CastProxy & Player.  Then the
Player destroys the AdManager.  The AdManager then triggers events
that the UI elements listen to.  Those events cause the UI elements to
invoke the destroyed CastProxy.

Therefore it is important to destroy the UI elements before the
CastProxy and Player.

This was discovered while working on an unrelated issue, by observing
errors logs from automated UI tests.  The logs contained exceptions
with stack traces.

Change-Id: I6c1d070255bb138967d5e733c0c7ec8dd16350b6