Fix race and exception in StreamingEngine
Playing content in our demo with my WiFi acting CRAZY, I found an
exception in StreamingEngine.  The bad connection seems to have
exacerbated and exposed a race condition that can be triggered when
the active stream is switched during an async appendBuffer() call.

The fix is to be careful not to assume that the mediaState object is
the same as it was after the async call.  If mediaState.stream has
changed, we should not manipulate mediaState.segmentIterator, which
may refer to a different Stream or may be null.

This bug does not affect any release versions.

Change-Id: I3d00009e78978cae95b47bb9f862fca4294b7b20