Use the module wrapper in debug builds
When we wrote the build scripts and module wrapper, we must have
assumed that module loaders would be used only in production, and
didn't consider that they may be required by an environment like react
even for debugging.

This changes the debug build so that the module wrapper is applied,
allowing react apps or other module-loader-based apps to switch
trivially to the debug build.

There is a slight difference, which is that non-exported classes are
not available through the wrapper. Private members of objects are
still accessible with their original names in the debug build with the
wrapper.

Closes #2465

Change-Id: Ife9374455303abaa1573dc5f763bc867a729770c