Fix ratechange events w/ src= playback
This fixes a missing listener for the ratechange event for src=
playbacks, and also fixes a small mistake in the MediaSource setup
that could have resulted in the ratechange event being triggered
before playRateController_ was created.

Closes #2488

Backported to v2.5.x

Note that v2.5.x does not have the playback rate feature in the UI,
nor the Player event for ratechange.

Change-Id: If85e489d681cfbb1ae6141b490d82264e1d932bd