Treat caption enabling as a text change.
Due to changes in when 'textchanged' events are sent out, the UI
was not moving the check mark off of 'off' when text was enabled, in
some situations. So this change just changes the 'texttrackvisibility'
event handler to also count as a language change, if text is being
made visible.
This is only a problem in v2.5.x, so this change will be temporary.

Closes #2484

Change-Id: I812b6950db0c209cff6f3ee0fcbfb911106f8c0a