Fix exception when enabling CEA text
This exception was thrown when we attempted to set up a
SegmentIterator on CEA tracks.  Because they do not have their own
segments, this is not possible.  The fix is to exclude CEA tracks from
this step.

Discovered while debugging #2395, but seems to be unrelated.

This bug does not affect any release versions.

Change-Id: I534b19139d88cc7ad2d6167611c04f400308a71b