Fix src= load promise callback on iOS Safari.
For whatever reason, on iOS Safari, the 'loadeddata' and
'loadedmetadata' events don't fire until the video actually starts
to play. That is to say, until the play button is pressed or the load
method is called (or the video is set to autoplay).
This changes the src= playback to call video.load.

Closes #2483

Change-Id: I49eb119a01d88073532496b5af062bbbcfcfa9fa