Fix missing EME polyfill in Cast receiver
This has not had an impact yet because we do not yet make use of the
encryption scheme queries that the polyfill provides.  But this
polyfill has been missing from the Cast receiver in all v2.5.x
releases since it was introduced.

The omission was first noticed when loading the Cast receiver in a
debugger in desktop Chrome.  This wouldn't normally be an issue with
most polyfills, but this one is outside of the Closure deps system.

Change-Id: I53685f5efe5555979993bf692d8a7cae31305ea9