Fix Shaka UI controls in Cast receiver
The Cast receiver does not ever change the controls between our UI and
the native controls, but the demo app does.  So in the demo app, we
never noticed that they do not seem to show up by default any more.
This fixes the UI library to show its own control by default, which in
turn fixes the Cast receiver app and any other app which did not use
setEnabledShakaControls().

Change-Id: I7889f1a7ec612bc7a260f5433316104d64ad4a26