Fix PlayRateController leak
We were leaking one PlayRateController instance per playback, with
each one running an infinite repeating timer that fired every 250 ms.

In a single run of our tests, this bug fix gets triggered 116 times.
That means we had, by the end of a test run, 116 timers with callbacks
every 250 ms.  Each would have to complete in 2.1 ms to avoid bogging
down the browser's JavaScript thread.

This bug could have had a significant performance impact in the field,
depending on usage patterns, and was part of v2.5.0 - v2.5.10.

I found this by running the Karma/Jasmine tests in a foreground
browser tab at http://localhost:9876/debug.html .  After the tests
completed, I hit the "pause" button in the JS debugger, and found that
this periodic timer was still running.

After the fix, pausing after the test run is complete does nothing.
There is no JavaScript executing beyond the end of the tests.

Change-Id: I87b402e733b1ccaa5e696d52d31cfe00ef12cfe0