Fix random test execution order
Jasmine 3+ defaults to random test order.  Since we own our own boot
file, we have to explicitly disable this if we want a predictable
order.  The unpredictable order has been uncovering some instability
and possible pollution between test suites lately, but before we can
address that, we need to stabilize our nightly test runs.  Then, when
hunting down these instabilities, we can explicitly use --random and
log the seed we used to get consistent failures once the failure is
identified.

This behavior was introduced to Jasmine in
https://github.com/jasmine/jasmine/commit/e31db20e and we unknowingly
adopted it in Shaka Player in
https://github.com/google/shaka-player/commit/b042a819 .

Bug: 154522761

Change-Id: I1b8b7b7f4e3d64746d61f019b1bd1d928d489f1b