Fix test order stability with xfilterDescribe
When running tests with --random --seed, the test order should be
stable.  However, xfilterDescribe (the version of filterDescribe which
temporarily disables tests) was not actually defining the tests
inside, which led to a change in the order of the tests run.

This makes xfilterDescribe define the tests, but disable them all,
which makes the shuffled order with --seed stable again.

Change-Id: I382f1f719a5b9f846bcb23761cc63f96abf17759