Generate externs from partially exported classes
In v2.5, we had some classes which were not exported, but which had
certain individual methods which were exported.  This is still true in
v2.6, except that we're using ES6 classes everywhere instead of mostly
using .prototype.

The extern generator had support for exporting individual .prototype
methods, but not for exporting individual class methods (unless the
class itself was exported already).

This adds support in the extern generator for these "partially
exported" classes, to correct missing externs in preparation for the
v2.6 launch.

Change-Id: I17ef377980ce771d74fab78afe26aa36ba3cce90