Suppress a bogus compiler error in StreamingEngine
This adds an annotation to suppress a bogus compiler error in
StreamingEngine.  In this method, we set updateTimer to null, then
await some async operation, then check to see if it's still null.  The
latest Closure Compiler thinks this is a useless check, but we know
better.  The updateTimer could have been created while the async
operation was in progress.

Issue #2528

Change-Id: I9a28b7dac1d7cb8bcfd836847507454030af52c6