Use brackets in VTTCue polyfill
These polyfills don't have the exact same type as the thing they are
polyfilling.  Instead of being constructors for VTTCue (which doesn't
exist if we are using the polyfill), they are functions that return
TextTrackCue (the base class of VTTCue and the closest thing we have
on some platforms).

Because the types don't match, the latest Closure Compiler complains.
To get around this, we use square brackets.

Issue #2528

Change-Id: Ie1e3f13dc51f2145f7d3c9d510b64d9a8d7b1f64