Fix UI keyboard Event types
The latest Closure Compiler is more picky about Event types, and
complained about the use of "key" on Event.  This updates the Event
types so that the compiler has the correct type info.

This also stops using "keyCode", which is deprecated, and only uses
the current "key" property, which is even supported on IE, and should
not create any compatibility issues.

Issue #2528

Backported to v2.5.x

Change-Id: Ic565772b1cc9597e358df015a73c40ac245edd6a