Fix demo usage of Storage API
The Storage API just changed to return an AbortableOperation instead
of a Promise.  The demo did not adopt this new API at the same time.

The runtime backward compatibility provided by the Storage API was
working, so the demo was not broken.  But we should be using the API
as defined.

This was caught by a compiler upgrade.

Issue #2528

Change-Id: I2d427ffb0400739c5d2de17a06c5fd79de3d39b2