Work around lack of error event type info in demo
The demo listens for shaka.Player.ErrorEvent, but it doesn't have type
information for that type.  This is because the type is defined only
in the docs, and not in the externs.  There is no actual class defined
for the event types.

The extern generator should be improved to handle this, but in the
meantime, we will use square brackets.

This was caught by a compiler upgrade.

Issue #2528

Change-Id: I55fc5fcc7f391f3155c1edc712064c5091fa363e