Fix compiler errors in invokeSpy
The invokeSpy utility is meant to work around an issue where the
compiler doesn't understand that jasmine.Spy is a callable function.
That workaround needed to be updated for the newest Closure Compiler
release.

That said, it's not clear why we should need this utility at all.  We
should be able to inform the compiler that this type is a callable
function.  This change also adds notes about possible future cleanup
and relevant issues filed against the compiler.

Issue #2528

Change-Id: I0e9a0e9821bd9a81fcf1349aaba6039ca59dbeeb