Add type info for test scheme metadata types
These metadata types were never defined, and the old compiler would
allow this.  The newer compiler is more strict, so this adds typedefs
for the metadata we use to generate manifests and streams for tests.

This also fixes a bug where we didn't correctly set the mimeType field
of the resulting streams.

This was caught by a compiler upgrade.

Issue #2528

Change-Id: Id16606b5c118793ee17dcf28942bf04d0ee22ba9